♻️ 제목 스타일과 동아리 소개 코드 리팩토링
